+++ app/Http/Controllers/UsersController.php
@@ -13,6 +13,7 @@ use View;
 use WA\Helpers\Traits\SetLimits;
 
 use WA\DataStore\User\User;
+use WA\DataStore\Role\Role;
 use WA\DataStore\User\UserTransformer;
 use WA\Repositories\User\UserInterface;
 
@@ -26,6 +27,11 @@ use WA\DataStore\Asset\Asset;
 use DB;
 use Cache;
 use Log;
+
+use JWTAuth;
+use Firebase\JWT\JWT;
+
+use WA\Events\UserCreatedEvent;
 
 /**
  * Users resource.
@@ -110,12 +116,34 @@ class UsersController extends FilteredApiController
             return response()->json($error)->setStatusCode($this->status_codes['badrequest']);
         }
 
+        // Deskpro JWT
+        $now = time();
+        $exp = $now + (60) * $_ENV['TOKEN_EXPIRES_IN'];
+
+        $deskproToken = array(
+            "jti"   => md5($now . rand()),
+            "iat"   => $now,
+            "exp"   => $exp,
+            "id"    => $user['id'],
+            "email" => $user['email'],
+            "name"  => $user['firstName'] . ' ' . $user['lastName']
+        );
+
+        $user->deskproJwt = JWT::encode($deskproToken, $_ENV['DESKPRO_JWT_SECRET']);
+
         $response = $this->response->item($user, new UserTransformer(), ['key' => $this->modelPlural]);
         $response = $this->applyMeta($response);
         return $response;
     }
 
-    public function store($id, Request $request) 
+    public function getJwtUserToken(Request $request)
+    {
+        $user = \Auth::user();
+        $token = JWTAuth::fromUser($user);
+        return array("jwttoken" => $token);
+    }
+
+    public function store($id, Request $request)
     {
         $success = true;
         $code = 'conflict';
@@ -127,6 +155,11 @@ class UsersController extends FilteredApiController
         if (!$this->isJsonCorrect($request, 'users')) {
             $error['errors']['json'] = Lang::get('messages.InvalidJson');
             return response()->json($error)->setStatusCode($this->status_codes['conflict']);
+        }
+
+        if(!$this->addFilterToTheRequest("store", $request)) {
+            $error['errors']['autofilter'] = Lang::get('messages.FilterErrorNotUser');
+            return response()->json($error)->setStatusCode($this->status_codes['notexists']);
         }
 
         DB::beginTransaction();
@@ -207,15 +240,15 @@ class UsersController extends FilteredApiController
                 }
             }
 
-            if (isset($dataRelationships['udls']) && $success) {
-                if (isset($dataRelationships['udls']['data'])) {
-                    $dataUdls = $this->parseJsonToArray($dataRelationships['udls']['data'], 'udls');
+            if (isset($dataRelationships['udlvalues']) && $success) {
+                if (isset($dataRelationships['udlvalues']['data'])) {
+                    $dataUdls = $this->parseJsonToArray($dataRelationships['udlvalues']['data'], 'udlvalues');
                     try {
                         $user->udlValues()->sync($dataUdls);
                     } catch (\Exception $e) {
                         $success = false;
-                        $error['errors']['udls'] = Lang::get('messages.NotOptionIncludeClass',
-                            ['class' => 'User', 'option' => 'updated', 'include' => 'Udls']);
+                        $error['errors']['udlvalues'] = Lang::get('messages.NotOptionIncludeClass',
+                            ['class' => 'User', 'option' => 'updated', 'include' => 'UdlValues']);
                         //$error['errors']['Message'] = $e->getMessage();
                     }
                 }
@@ -397,29 +430,46 @@ class UsersController extends FilteredApiController
             return response()->json($error)->setStatusCode($this->status_codes['conflict']);
         }
 
+        $data = $request->all()['data'];
+        $data = $this->addRelationships($data);
+
+        if(!$this->addFilterToTheRequest("create", $data)) {
+            $error['errors']['autofilter'] = Lang::get('messages.FilterErrorNotUser');
+            return response()->json($error)->setStatusCode($this->status_codes['notexists']);
+        }
+
         DB::beginTransaction();
 
         /*
          * Now we can create the User.
          */
         try {
-            $data = $request->all()['data'];
-
             if($this->userInterface->byEmail($data['attributes']['email'])['id'] > 0) {
                 $error['errors']['User'] = 'The User can not be created, there are other user with the same email.';
                 return response()->json($error)->setStatusCode(409);
             }
 
             $user = $this->userInterface->create($data['attributes']);
+
             if(!$user){
                 $error['errors']['users'] = 'The User has not been created, some data information is wrong, may be the Email.';
                 return response()->json($error)->setStatusCode(409);
             }
+
+            $userRole = Role::where('name', 'user')->first();
+            if (isset($userRole)) {
+                $user->roles()->sync([$userRole->id]);
+            } else {
+                $success = false;
+                $error['errors']['users'] = Lang::get('messages.NotOptionIncludeClass',
+                    ['class' => 'User', 'option' => 'created', 'include' => 'Role']);
+            }           
+
         } catch (\Exception $e) {
             $success = false;
             $error['errors']['users'] = Lang::get('messages.NotOptionIncludeClass',
                 ['class' => 'User', 'option' => 'created', 'include' => '']);
-            //$error['errors']['Message'] = $e->getMessage();
+            $error['errors']['Message'] = $e->getMessage();
         }
 
         /*
@@ -441,7 +491,7 @@ class UsersController extends FilteredApiController
                     }
                 }
             }
-
+/*
             if (isset($dataRelationships['roles']) && $success) {
                 if (isset($dataRelationships['roles']['data'])) {
                     $dataRoles = $this->parseJsonToArray($dataRelationships['roles']['data'], 'roles');
@@ -455,16 +505,16 @@ class UsersController extends FilteredApiController
                     }
                 }
             }
-
-            if (isset($dataRelationships['udls']) && $success) {
-                if (isset($dataRelationships['udls']['data'])) {
-                    $dataUdls = $this->parseJsonToArray($dataRelationships['udls']['data'], 'udls');
+*/
+            if (isset($dataRelationships['udlvalues']) && $success) {
+                if (isset($dataRelationships['udlvalues']['data'])) {
+                    $dataUdls = $this->parseJsonToArray($dataRelationships['udlvalues']['data'], 'udlvalues');
                     try {
                         $user->udlValues()->sync($dataUdls);
                     } catch (\Exception $e) {
                         $success = false;
-                        $error['errors']['udls'] = Lang::get('messages.NotOptionIncludeClass',
-                            ['class' => 'User', 'option' => 'created', 'include' => 'Udls']);
+                        $error['errors']['udlvalues'] = Lang::get('messages.NotOptionIncludeClass',
+                            ['class' => 'User', 'option' => 'created', 'include' => 'UdlValues']);
                         //$error['errors']['Message'] = $e->getMessage();
                     }
                 }
@@ -518,23 +568,15 @@ class UsersController extends FilteredApiController
 
                     try {
                         $interfaceC = app()->make('WA\Repositories\Content\ContentInterface');
+                        
+                        foreach ($data as $content) {
+                            $content['owner_id'] = $user->id;
+                            $interfaceC->create($content);
+                        }
                     } catch (\Exception $e) {
                         $success = false;                        
-                        $error['errors']['contents'] = Lang::get('messages.NotOptionIncludeClass', ['class' => 'User', 'option' => 'created', 'include' => 'contents']);
+                        $error['errors']['contents'] = Lang::get('messages.NotOptionIncludeClass', ['class' => 'User', 'option' => 'created', 'include' => 'Contents']);
                         //$error['errors']['Message'] = $e->getMessage();
-                    }
-
-                    if ($success) {
-                        try {                           
-                            foreach ($data as $content) {
-                                $content['owner_id'] = $user->id;
-                                $interfaceC->create($content);
-                            }
-                        } catch (\Exception $e) {
-                            $success = false;
-                            $error['errors']['contents'] = Lang::get('messages.NotOptionIncludeClass', ['class' => 'user', 'option' => 'created', 'include' => 'contents']);
-                            //$error['errors']['Message'] = $e->getMessage();
-                        }
                     }
                 }
             }
@@ -577,23 +619,30 @@ class UsersController extends FilteredApiController
                 'redirectPath' => $redirectPath,
             ];
 
-            $mail = Mail::send('emails.auth.register', $data, function ($m) use ($user) {
-                $m->from(env('MAIL_FROM_ADDRESS'), 'Wireless Analytics');
-                $m->to($user->email)->subject('New User '.$user->username.' !');
-            });
+//            $mail = Mail::send('emails.auth.register', $data, function ($m) use ($user) {
+//                $m->from(env('MAIL_FROM_ADDRESS'), 'Wireless Analytics');
+//                $m->to($user->email)->subject('New User '.$user->username.' !');
+//            });
 
             DB::commit();
             Cache::put('user_email_'.$code, $user->identification, 60);
             Cache::put('user_code_'.$user->identification, $code, 60);
+
             return $this->response()->item($user, new UserTransformer(), ['key' => 'users'])
-                ->setStatusCode($this->status_codes['created']);    
+                ->setStatusCode($this->status_codes['created']);
         } else {
             DB::rollBack();
             return response()->json($error)->setStatusCode($this->status_codes['conflict']);
         }
     }
 
-    public function delete($id) {
+    public function delete($id)
+    {
+        if(!$this->addFilterToTheRequest("delete", null)) {
+            $error['errors']['autofilter'] = Lang::get('messages.FilterErrorNotUser');
+            return response()->json($error)->setStatusCode($this->status_codes['notexists']);
+        }
+
         $user = User::find($id);
         if ($user <> null) {
             $this->userInterface->deleteById($id);
@@ -610,26 +659,6 @@ class UsersController extends FilteredApiController
             return response()->json($error)->setStatusCode($this->status_codes['conflict']);
         }
     }
-
-    /*public function getLoggedInUser(Request $request)
-    {
-        $user = Auth::user();
-
-        if ($user === null) {
-            $error['errors']['scopes'] = 'There\'s no user authenticated';
-            return response()->json($error)->setStatusCode($this->status_codes['notexists']);
-        }
-
-        $transformer = $user->getTransformer();
-
-        if (!$this->includesAreCorrect($request, $transformer)) {
-            $error['errors']['getIncludes'] = Lang::get('messages.NotExistInclude');
-            return response()->json($error)->setStatusCode($this->status_codes['badrequest']);
-        }
-
-        $response = $this->response->item($user, $transformer, ['key' => 'users']);
-        return $response;
-    }*/
 
     private function transformToJson($array) {
         $list = [];
@@ -651,7 +680,7 @@ class UsersController extends FilteredApiController
         $udlV = $user->udlValues;
         $arrayAux = [];
         foreach ($udlV as $uv) {
-            $aux['udlName'] = $uv->udl->name;
+            $aux['udlName'] = $uv->udls->name;
             $aux['udlValue'] = $uv->name;
             array_push($arrayAux, $aux);
         }
